Extrae noticias de Ciper Chile

Esta funcion se conecta a la API de Ciper para descargar noticias.

Usage

extraer_noticias_max_res_ciper(search_query = NULL, max_results = NULL)

Arguments

search_query

El termino de busqueda (opcional). Si es NULL, extrae todos los articulos.

max_results

El numero maximo de noticias a extraer.

Value

Un dataframe con las noticias extraidas, estandarizado al formato de datamedios.

Examples

# \donttest{
# Extraer los ultimos 5 articulos con una busqueda
noticias_ciper <- extraer_noticias_max_res_ciper("corrupcion", max_results = 5)
#> Iniciando extraccion desde Ciper para la busqueda: corrupcion
#> Total de resultados encontrados en Ciper: 354
#>   |                                                                              |                                                                      |   0%
#> Iniciando scrapeo para 5 articulos...
#> Scrapeo en paralelo completado.

# Extraer los ultimos 10 articulos sin busqueda
ultimos_ciper <- extraer_noticias_max_res_ciper(max_results = 10)
#> Iniciando extraccion de todos los articulos desde Ciper (sin busqueda).
#> Total de resultados encontrados en Ciper: 7234
#>   |                                                                              |                                                                      |   0%
#> Iniciando scrapeo para 10 articulos...
#> Scrapeo en paralelo completado.
# }