Funcion de nube de palabras

Esta funcion permite realizar una nube de palabras con las palabras más frecuentes del corpus de noticias.

Usage

word_cloud(datos, max_words, stop_words = NULL)

Arguments

datos

data frame que incluye la columna contenido_limpio.

max_words

Cantidad maxima de palabras que apareceran en la nube.

stop_words

Definir las palabras que seran ignoradas en la visualizacion. Debe ser un vector de carácteres.

Value

Una nube de palabras con las palabras mas frecuentes.

Examples

if (FALSE) { # \dontrun{
datos <- extraer_noticias_fecha("Boric",
"2025-03-01",
"2025-04-01",
fuentes="bbcl",
subir_a_bd = FALSE)
datos_proc <- limpieza_notas(datos)
word_cloud(datos_proc, max_words = 50, stop_words = c("es", "la"))
} # }