Generar una tabla estilizada con las palabras mas frecuentes

Esta funcion procesa la columna 'contenido_limpio' de un dataframe, tokeniza el texto, cuenta la frecuencia de cada palabra y genera una tabla con las palabras mas frecuentes.

tabla_frecuencia_palabras(datos, max_words, stop_words = NULL)

Arguments

datos

Data frame que contiene la columna 'contenido_limpio'.

max_words

Numero maximo de palabras que se mostraran en la tabla.

stop_words

Vector opcional de palabras que se deben excluir del conteo.

Value

Una tabla con las palabras mas frecuentes.

Examples

datos <- extraer_noticias_fecha("cambio climatico", "2024-01-01","2025-01-01", subir_a_bd = FALSE)
#> [1] "Total de resultados posibles: 5656"
#> [1] "Noticia mas reciente disponible es de la fecha: 2025-01-24 13:41:22"
#> [1] "No hay mas noticias dentro del rango de fechas. Terminando la busqueda."
#> [1] "Total de noticias encontradas en el rango de fechas: 547"
datos <- extraccion_parrafos(datos)
datos <- limpieza_notas(datos)
#> [1] 547
#> [1] 0
#> El numero total de resultados obtenidos es: 0 
tabla_frecuencia_palabras(datos, max_words = 5, stop_words = c("el", "de"))